Change filters to use in and out streams instead of strings
Using strings requires the whole body to be in memory at one time. If
the body is too big heap exhaustion occurs. To decouple filters from
http protocol, the body is read into a smart buffer before going
through the filter; this extra processing is avoided in the filters
are not used (nil).